Fix "add item" overflowing the sort_order column
sort_order was declared as a 32-bit integer but new items were
stamped with Date.now() (milliseconds), which exceeds that range and
made every add fail with Postgres error 22003. Uses a seconds-based
timestamp instead and widens the column to bigint so it can't
overflow again.

+++ supabase/schema.sql
@@ -7,7 +7,7 @@ create table if not exists menu_items (
   category_label text not null,
   name text not null,
   price text not null,
-  sort_order integer not null,
+  sort_order bigint not null,
   is_custom boolean not null default false,
   is_deleted boolean not null default false,
   last_edited_by text,
